+++ projects/_01_ingest/cqc_api/fargate/validate_delta_locations_api_cleaned.py
@@ -3,10 +3,9 @@ import sys
 import pointblank as pb
 import polars as pl
 
-from polars_utils import utils
+from polars_utils import raw_data_adjustments, utils
 from polars_utils.expressions import has_value, str_length_cols
 from polars_utils.logger import get_logger
-from polars_utils.raw_data_adjustments import is_valid_location
 from polars_utils.validation import actions as vl
 from polars_utils.validation.constants import GLOBAL_ACTIONS, GLOBAL_THRESHOLDS
 from utils.column_names.cleaned_data_files.cqc_location_cleaned import (
@@ -17,11 +16,6 @@ from utils.column_names.raw_data_files.cqc_location_api_columns import (
     NewCqcLocationApiColumns as CQCL,
 )
 from utils.column_names.validation_table_columns import Validation
-from utils.column_values.categorical_column_values import (
-    LocationType,
-    RegistrationStatus,
-    Services,
-)
 from utils.column_values.categorical_columns_by_dataset import (
     LocationsApiCleanedCategoricalValues as CatValues,
 )
@@ -61,6 +55,7 @@ def main(
         f"s3://{bucket_name}/{compare_path}",
         selected_columns=compare_columns_to_import,
     )
+    expected_row_count = expected_size(compare_df)
 
     validation = (
         pb.Validate(
@@ -71,7 +66,10 @@ def main(
             actions=GLOBAL_ACTIONS,
         )
         # dataset size
-        .row_count_match(expected_size(compare_df))
+        .row_count_match(
+            expected_row_count,
+            brief=f"Cleaned file has {source_df.height} rows but expecting {expected_row_count} rows",
+        )
         # complete columns
         .col_vals_not_null(
             [
@@ -82,6 +80,7 @@ def main(
                 CQCLClean.registration_status,
                 CQCLClean.imputed_registration_date,
                 CQCLClean.name,
+                CQCLClean.type,
             ]
         )
         # index columns
@@ -91,8 +90,9 @@ def main(
                 CQCLClean.cqc_location_import_date,
             ],
         )
+        # greater than or equal to
+        .col_vals_ge(CQCLClean.number_of_beds, 0, na_pass=True)
         # between (inclusive)
-        .col_vals_between(CQCLClean.number_of_beds, 0, 500, na_pass=True)
         .col_vals_between(Validation.location_id_length, 3, 14)
         .col_vals_between(Validation.provider_id_length, 3, 14)
         # categorical
@@ -150,6 +150,7 @@ def main(
 
 def expected_size(df: pl.DataFrame) -> int:
     gac_services = pl.col(CQCL.gac_service_types)
+
     cleaned_df = df.with_columns(
         # nullify empty lists to avoid index out of bounds error
         pl.when(gac_services.list.len() > 0).then(gac_services),
@@ -157,6 +158,9 @@ def expected_size(df: pl.DataFrame) -> int:
         # TODO: remove regulated_activities
         has_value(df, CQCL.regulated_activities, CQCL.location_id),
         has_value(df, CQCL.provider_id, CQCL.location_id),
+        has_value(df, CQCL.registration_status, CQCL.location_id),
+        has_value(df, CQCL.type, CQCL.location_id),
+        raw_data_adjustments.is_valid_location(),
     )
     logger.info(f"Expected size {cleaned_df.height}")
     return cleaned_df.height