+++ projects/_01_ingest/cqc_api/tests/fargate/test_validate_delta_locations_api_cleaned.py
@@ -52,6 +52,7 @@ class ValidateLocationsRawTests(unittest.TestCase):
             {
                 "assertion_type": "row_count_match",
                 "all_passed": False,
+                "brief": "Cleaned file has 3 rows but expecting 0 rows",
             },
             report_json[0],
         )
@@ -77,16 +78,16 @@ class ValidateLocationsRawTests(unittest.TestCase):
                 "column": ["locationId", "cqc_location_import_date"],
                 "all_passed": True,
             },
-            report_json[8],
+            report_json[9],
         )
         self.assertDictContainsSubset(
             {
-                "assertion_type": "col_vals_between",
+                "assertion_type": "col_vals_ge",
                 "column": "numberOfBeds",
-                "all_passed": True,
+                "all_passed": False,
                 "na_pass": True,
             },
-            report_json[9],
+            report_json[10],
         )
         self.assertDictContainsSubset(
             {
@@ -95,7 +96,7 @@ class ValidateLocationsRawTests(unittest.TestCase):
                 "all_passed": False,
                 "n_failed": 2,
             },
-            report_json[10],
+            report_json[11],
         )
         self.assertDictContainsSubset(
             {
@@ -104,7 +105,7 @@ class ValidateLocationsRawTests(unittest.TestCase):
                 "all_passed": False,
                 "n_failed": 2,
             },
-            report_json[11],
+            report_json[12],
         )
         self.assertDictContainsSubset(
             {
@@ -116,7 +117,7 @@ class ValidateLocationsRawTests(unittest.TestCase):
                     "Independent",
                 ],
             },
-            report_json[12],
+            report_json[13],
         )
         self.assertDictContainsSubset(
             {
@@ -125,7 +126,7 @@ class ValidateLocationsRawTests(unittest.TestCase):
                 "all_passed": True,
                 "values": ["Y", "N", None],
             },
-            report_json[14],
+            report_json[15],
         )
         self.assertDictContainsSubset(
             {
@@ -133,7 +134,7 @@ class ValidateLocationsRawTests(unittest.TestCase):
                 "all_passed": True,
                 "brief": "dormancy needs to be null, or one of ['Y', 'N']",
             },
-            report_json[18],
+            report_json[19],
         )